fix: More defensive code when trying to drop tables

+++ migration/Horde/Perms/2_horde_perms_upgrade_autoincrement.php
@@ -4,9 +4,8 @@ class HordePermsUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('horde_perms', 'perm_id', 'autoincrementKey');
-        try {
+        if (in_array('horde_perms_seq', $this->tables())) {
             $this->dropTable('horde_perms_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
